Moving createOffer to its proper controller
